refactor(kubernetes): rename kubernetes client
