+++ kubernetes/kubernetesversions.go
@@ -1,4 +1,4 @@
-package kubernetesclient
+package kubernetes
 
 import (
 	"context"
